Fix BigInt serialization issues in admin properties and property details
- Fixed admin properties list 500 error with proper Prisma import and BigInt serialization
- Fixed property detail view 500 error with same BigInt fixes
- Fixed Cloudinary signature generation for video uploads
- Updated price validation to support ₦900M properties
- All admin features now working with BigInt price fields

+++ src/app/admin/properties/create/page.tsx
@@ -340,7 +340,7 @@ export default function CreatePropertyPage() {
         landmark: '',
 
         // Pricing
-        price: BigInt(parseInt(formData.price.replace(/,/g, '')) * 100), // Convert to kobo as BigInt
+        price: parseInt(formData.price.replace(/,/g, '')) * 100, // Convert to kobo as number
         currency: 'NGN' as const,
         isNegotiable: true,
